Pdf_Analyser/requirements.txt
phidata
python-dotenv
yfinance
duckduckgo-search
packaging
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf